[Portal] Make the applications sections reusable outside their screens
Groundwork for comparing layouts of this surface without cloning any of
it: whichever arrangement wins, the pieces it arranges should have one
definition.

The static client list moves out of ApplicationsConfigurationScreen
into ApplicationsListSection, along with the form state it edits, and
the block-action quota helper into util/usageLimit. Each mechanism
screen gives up its ScreenContent and page title to its own wrapper and
exports its content component and form-state functions, so the cards
can also be rendered by a layout that supplies its own page chrome.

Structure only -- every component still renders in the same order with
the same props, and no error binding, dialog or save semantics moved.

ref DEV-3838
