+++ portal/src/graphql/portal/ApplicationsConfigurationScreen.tsx
@@ -1,24 +1,10 @@
-import React, { useCallback, useContext, useMemo, useState } from "react";
-import cn from "classnames";
-import {
-  DropdownMenu,
-  Heading,
-  IconButton as RadixIconButton,
-  Text,
-} from "@radix-ui/themes";
-import { DotsVerticalIcon } from "@radix-ui/react-icons";
+import React, { useCallback, useContext, useMemo } from "react";
+import { Heading } from "@radix-ui/themes";
 import { Context, FormattedMessage } from "../../intl";
-import { useNavigate, useParams } from "react-router-dom";
-import { produce } from "immer";
+import { useParams } from "react-router-dom";
 
 import ShowError from "../../ShowError";
 import ShowLoading from "../../ShowLoading";
-import {
-  OAuthClientConfig,
-  PortalAPIAppConfig,
-  StandingUsageLimitConfig,
-} from "../../types";
-import { clearEmptyObject } from "../../util/misc";
 import {
   AppConfigFormModel,
   useAppConfigForm,
@@ -26,192 +12,17 @@ import {
 import styles from "./ApplicationsConfigurationScreen.module.css";
 import ScreenContent from "../../ScreenContent";
 import { useAppFeatureConfigQuery } from "./query/appFeatureConfigQuery";
-import { getApplicationTypeMessageID } from "./EditOAuthClientForm";
-import { findFramework } from "./CreateOAuthClientScreen/frameworks";
 import FormContainer from "../../FormContainer";
-import { PrimaryButton } from "../../components/v2/Button/PrimaryButton/PrimaryButton";
-import { CardTable } from "../../components/v2/CardTable/CardTable";
-import { ConfirmationDialog } from "../../components/v2/ConfirmationDialog/ConfirmationDialog";
-import { CopyIconButton } from "../../components/v2/CopyIconButton/CopyIconButton";
-import { FeatureDisabledCallout } from "../../components/v2/FeatureDisabledCallout/FeatureDisabledCallout";
 import { OverflowTabs } from "../../components/v2/OverflowTabs/OverflowTabs";
-import { Tooltip } from "../../components/v2/Tooltip/Tooltip";
-import { useOAuthClientForm } from "../../hook/useOAuthClientForm";
 import { usePivotNavigation } from "../../hook/usePivot";
-import { getNextPlan } from "../../util/plan";
-import { RolesAndGroupsEmptyView } from "../../components/roles-and-groups/empty-view/RolesAndGroupsEmptyView";
+import { smallestBlockQuota } from "../../util/usageLimit";
 import { DynamicClientsTab } from "../../components/dynamic-clients/DynamicClientsTab";
-
-// Only the static client list. CIMD and DCR each own their own nav-level
-// screen and their own form, so neither mechanism's fields belong here: a
-// save from this screen must never carry an edit made on one of them.
-export interface FormState {
-  clients: OAuthClientConfig[];
-}
-
-function constructFormState(config: PortalAPIAppConfig): FormState {
-  return {
-    clients: config.oauth?.clients ?? [],
-  };
-}
-
-function constructConfig(
-  config: PortalAPIAppConfig,
-  _initialState: FormState,
-  currentState: FormState
-): PortalAPIAppConfig {
-  const [newConfig, _] = produce(
-    [config, currentState],
-    ([config, currentState]) => {
-      config.oauth ??= {};
-      config.oauth.clients = currentState.clients;
-
-      clearEmptyObject(config);
-    }
-  );
-  return newConfig;
-}
-
-// A standing usage limit can carry several entries; only an `action: block`
-// entry stops anything, and the smallest of those is the effective cap.
-// Returns null when the plan does not cap this usage at all.
-function smallestBlockQuota(
-  limits: StandingUsageLimitConfig[] | undefined
-): number | null {
-  const blockQuotas = (limits ?? [])
-    .filter((limit) => limit.action === "block")
-    .map((limit) => limit.quota)
-    .filter((quota): quota is number => quota != null);
-  if (blockQuotas.length === 0) {
-    return null;
-  }
-  return Math.min(...blockQuotas);
-}
-
-function stopPropagation(e: React.SyntheticEvent) {
-  e.stopPropagation();
-}
-
-interface ClientRowProps {
-  client: OAuthClientConfig;
-  onDeleteClick: (clientID: string) => void;
-}
-
-const ClientRow: React.VFC<ClientRowProps> = function ClientRow(props) {
-  const { client, onDeleteClick } = props;
-  const { renderToString } = useContext(Context);
-  const { appID } = useParams() as { appID: string };
-  const navigate = useNavigate();
-
-  const editPath = `/project/${appID}/configuration/apps/${encodeURIComponent(
-    client.client_id
-  )}/edit`;
-
-  const framework = findFramework(client.x_framework);
-  const fallbackIcon =
-    client.x_application_type === "m2m" ? "server" : "app-window";
-  const iconName = framework?.iconName ?? fallbackIcon;
-
-  const onRowClick = useCallback(() => {
-    navigate(editPath);
-  }, [navigate, editPath]);
-
-  const onRowKeyDown = useCallback(
-    (e: React.KeyboardEvent<HTMLDivElement>) => {
-      // Only navigate when the row itself is focused; Enter/Space on the
-      // copy button or the actions menu must not trigger navigation.
-      if (e.target !== e.currentTarget) {
-        return;
-      }
-      if (e.key === "Enter" || e.key === " ") {
-        e.preventDefault();
-        navigate(editPath);
-      }
-    },
-    [navigate, editPath]
-  );
-
-  return (
-    <CardTable.Row
-      className={styles.clientRow}
-      role="button"
-      tabIndex={0}
-      onClick={onRowClick}
-      onKeyDown={onRowKeyDown}
-    >
-      <CardTable.Cell className={styles.colName}>
-        <div className={styles.clientIconWrap}>
-          <i
-            className={cn("ti", `ti-${iconName}`, styles.clientIcon)}
-            aria-hidden={true}
-          />
-        </div>
-        <div className={styles.clientNameBlock}>
-          <Text size="2" className={styles.clientName}>
-            {client.name ?? ""}
-          </Text>
-          <Text size="1" className={styles.clientSubtitle}>
-            <FormattedMessage
-              id={getApplicationTypeMessageID(client.x_application_type)}
-            />
-            {framework != null ? (
-              <>
-                {" · "}
-                <FormattedMessage id={framework.displayNameMessageId} />
-              </>
-            ) : null}
-          </Text>
-          <div className={styles.compactClientId} onClick={stopPropagation}>
-            <Text size="1" className={styles.clientIdText}>
-              {client.client_id}
-            </Text>
-            <CopyIconButton textToCopy={client.client_id} />
-          </div>
-        </div>
-      </CardTable.Cell>
-      <CardTable.Cell className={styles.colClientId} onClick={stopPropagation}>
-        <Text size="2" className={styles.clientIdText}>
-          {client.client_id}
-        </Text>
-        <CopyIconButton textToCopy={client.client_id} />
-      </CardTable.Cell>
-      <CardTable.Cell className={styles.colActions} onClick={stopPropagation}>
-        <DropdownMenu.Root>
-          <DropdownMenu.Trigger>
-            <RadixIconButton
-              className={styles.rowActionsButton}
-              variant="soft"
-              color="gray"
-              size="2"
-              aria-label={renderToString(
-                "ApplicationsConfigurationScreen.client-list.row-actions"
-              )}
-            >
-              <DotsVerticalIcon width="1rem" height="1rem" />
-            </RadixIconButton>
-          </DropdownMenu.Trigger>
-          <DropdownMenu.Content align="end">
-            <DropdownMenu.Item
-              onSelect={() => {
-                navigate(editPath);
-              }}
-            >
-              <FormattedMessage id="edit" />
-            </DropdownMenu.Item>
-            <DropdownMenu.Item
-              color="red"
-              onSelect={() => {
-                onDeleteClick(client.client_id);
-              }}
-            >
-              <FormattedMessage id="ApplicationsConfigurationScreen.delete-client.label" />
-            </DropdownMenu.Item>
-          </DropdownMenu.Content>
-        </DropdownMenu.Root>
-      </CardTable.Cell>
-    </CardTable.Row>
-  );
-};
+import {
+  ApplicationsListSection,
+  FormState,
+  constructConfig,
+  constructFormState,
+} from "./ApplicationsListSection";
 
 // The dynamic-clients tab holds only what CIMD and DCR share; configuring
 // either mechanism happens on its own nav-level screen.
@@ -238,7 +49,7 @@ interface OAuthClientConfigurationContentProps {
 const OAuthClientConfigurationContent: React.VFC<OAuthClientConfigurationContentProps> =
   function OAuthClientConfigurationContent(props) {
     const {
-      form: { state, reload },
+      form,
       planName,
       oauthClientsHardMaximum,
       oauthClientsSoftMaximum,
@@ -247,81 +58,7 @@ const OAuthClientConfigurationContent: React.VFC<OAuthClientConfigurationContent
       dcrClientQuota,
       cimdClientQuota,
     } = props;
-    const navigate = useNavigate();
     const { renderToString } = useContext(Context);
-    const { appID } = useParams() as { appID: string };
-
-    const deleteForm = useOAuthClientForm(appID, null);
-
-    const [isRemoveDialogVisible, setIsRemoveDialogVisible] = useState(false);
-
-    const hardLimitReached = useMemo(() => {
-      if (oauthClientsHardMaximum == null) {
-        return false;
-      }
-      return state.clients.length >= oauthClientsHardMaximum;
-    }, [oauthClientsHardMaximum, state.clients.length]);
-
-    const displayedClientMaximum = useMemo<number | undefined>(() => {
-      return oauthClientsSoftMaximum ?? oauthClientsHardMaximum;
-    }, [oauthClientsHardMaximum, oauthClientsSoftMaximum]);
-
-    const goToCreateApp = useCallback(() => {
-      navigate(`/project/${appID}/configuration/apps/add`);
-    }, [appID, navigate]);
-
-    const showDialogAndSetRemoveClientByID = useCallback(
-      (clientID: string) => {
-        deleteForm.setState((state) => ({
-          ...state,
-          removeClientByID: clientID,
-        }));
-        setIsRemoveDialogVisible(true);
-      },
-      [deleteForm, setIsRemoveDialogVisible]
-    );
-
-    const dismissDialogAndResetRemoveClientByID = useCallback(() => {
-      setIsRemoveDialogVisible(false);
-      deleteForm.setState((state) => {
-        return {
-          ...state,
-          removeClientByID: undefined,
-        };
-      });
-    }, [deleteForm, setIsRemoveDialogVisible]);
-
-    const onRemoveDialogOpenChange = useCallback(
-      (open: boolean) => {
-        if (!open && !deleteForm.isUpdating) {
-          dismissDialogAndResetRemoveClientByID();
-        }
-      },
-      [deleteForm.isUpdating, dismissDialogAndResetRemoveClientByID]
-    );
-
-    const onConfirmRemove = useCallback(() => {
-      deleteForm.save().then(
-        () => {
-          dismissDialogAndResetRemoveClientByID();
-          reload();
-        },
-        () => {
-          dismissDialogAndResetRemoveClientByID();
-        }
-      );
-    }, [deleteForm, reload, dismissDialogAndResetRemoveClientByID]);
-
-    const canUpgradePlan = useMemo(() => {
-      return getNextPlan(planName ?? "") != null;
-    }, [planName]);
-
-    const displayMaximumWarning = useMemo(() => {
-      if (displayedClientMaximum == null) {
-        return false;
-      }
-      return state.clients.length >= displayedClientMaximum;
-    }, [state, displayedClientMaximum]);
 
     const onTabChange = useCallback(
       (value: string) => {
@@ -349,13 +86,9 @@ const OAuthClientConfigurationContent: React.VFC<OAuthClientConfigurationContent
     );
 
     const cimdEnabled =
-      props.form.effectiveConfig.oauth?.client_id_metadata_document?.enabled ??
-      false;
+      form.effectiveConfig.oauth?.client_id_metadata_document?.enabled ?? false;
     const dcrEnabled =
-      props.form.effectiveConfig.oauth?.dynamic_client_registration?.enabled ??
-      false;
-
-    const isEmpty = state.clients.length === 0;
+      form.effectiveConfig.oauth?.dynamic_client_registration?.enabled ?? false;
 
     return (
       <ScreenContent layout="list">
@@ -379,111 +112,14 @@ const OAuthClientConfigurationContent: React.VFC<OAuthClientConfigurationContent
               dcrEnabled={dcrEnabled}
             />
           </div>
-        ) : isEmpty ? (
-          <div className={cn(styles.widget, styles.emptyState)}>
-            <RolesAndGroupsEmptyView
-              icon={
-                <span className={styles.emptyStateIconWrap}>
-                  <i
-                    className={cn("ti", "ti-apps", styles.emptyStateIcon)}
-                    aria-hidden={true}
-                  />
-                </span>
-              }
-              title={
-                <FormattedMessage id="ApplicationsConfigurationScreen.empty-state.title" />
-              }
-              description={
-                <FormattedMessage id="ApplicationsConfigurationScreen.empty-state.description" />
-              }
-              button={
-                <RolesAndGroupsEmptyView.CreateButton
-                  onClick={goToCreateApp}
-                  text={
-                    <FormattedMessage id="ApplicationsConfigurationScreen.add-client-button" />
-                  }
-                />
-              }
-            />
-          </div>
         ) : (
-          <>
-            <div className={cn(styles.widget, styles.listHeader)}>
-              <Text as="p" size="2" className={styles.pageDescription}>
-                <FormattedMessage id="ApplicationsConfigurationScreen.description" />
-              </Text>
-              <Tooltip
-                content={
-                  <FormattedMessage
-                    id="ApplicationsConfigurationScreen.add-client-button.hard-limit-tooltip"
-                    values={{ maximum: oauthClientsHardMaximum ?? 0 }}
-                  />
-                }
-                disabled={!hardLimitReached}
-              >
-                {/* The tooltip must still fire when the button is disabled,
-                    so it anchors on a wrapper span instead of the button
-                    itself. */}
-                <span>
-                  <PrimaryButton
-                    size="2"
-                    text={
-                      <FormattedMessage id="ApplicationsConfigurationScreen.add-client-button" />
-                    }
-                    onClick={goToCreateApp}
-                    disabled={hardLimitReached}
-                  />
-                </span>
-              </Tooltip>
-            </div>
-            <div className={cn(styles.widget, styles.listSection)}>
-              {displayMaximumWarning ? (
-                <FeatureDisabledCallout
-                  messageID={
-                    canUpgradePlan
-                      ? "FeatureConfig.oauth-clients.maximum.upgrade"
-                      : "FeatureConfig.oauth-clients.maximum.contact-us"
-                  }
-                  messageValues={{ maximum: displayedClientMaximum! }}
-                />
-              ) : null}
-              <CardTable>
-                <CardTable.Header>
-                  <CardTable.HeaderCell className={styles.colName}>
-                    <FormattedMessage id="ApplicationsConfigurationScreen.client-list.name" />
-                  </CardTable.HeaderCell>
-                  <CardTable.HeaderCell className={styles.colClientId}>
-                    <FormattedMessage id="ApplicationsConfigurationScreen.client-list.client-id" />
-                  </CardTable.HeaderCell>
-                  <CardTable.HeaderCell className={styles.colActions} />
-                </CardTable.Header>
-                {state.clients.map((client) => (
-                  <ClientRow
-                    key={client.client_id}
-                    client={client}
-                    onDeleteClick={showDialogAndSetRemoveClientByID}
-                  />
-                ))}
-              </CardTable>
-            </div>
-          </>
+          <ApplicationsListSection
+            form={form}
+            planName={planName}
+            oauthClientsHardMaximum={oauthClientsHardMaximum}
+            oauthClientsSoftMaximum={oauthClientsSoftMaximum}
+          />
         )}
-        <ConfirmationDialog
-          open={isRemoveDialogVisible}
-          onOpenChange={onRemoveDialogOpenChange}
-          title={
-            <FormattedMessage id="ApplicationsConfigurationScreen.delete-client-dialog.title" />
-          }
-          description={
-            <FormattedMessage id="ApplicationsConfigurationScreen.delete-client-dialog.description" />
-          }
-          confirmText={<FormattedMessage id="confirm" />}
-          cancelText={<FormattedMessage id="cancel" />}
-          onConfirm={onConfirmRemove}
-          onCancel={dismissDialogAndResetRemoveClientByID}
-          loading={deleteForm.isUpdating}
-          confirmColor="red"
-        />
       </ScreenContent>
     );
   };
