+++ portal/src/graphql/portal/ApplicationsListSection.module.css
@@ -1,0 +1,123 @@
+.widget {
+  grid-column: 1 / span 12;
+  @apply tablet:col-span-full;
+}
+
+.pageDescription {
+  margin: 0;
+  color: var(--gray-a11);
+}
+
+.listSection {
+  @apply flex flex-col gap-4;
+}
+
+.emptyState {
+  @apply py-12;
+}
+
+/* The list-row blue icon tile, scaled to the empty view's 2rem icon slot. */
+.emptyStateIconWrap {
+  width: 2rem;
+  height: 2rem;
+  border-radius: 8px;
+  background: #eef4ff;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.emptyStateIcon {
+  font-size: 20px;
+  line-height: 1;
+  color: #0969da;
+}
+
+.colName {
+  @apply min-w-0 gap-3 overflow-hidden;
+  flex: 2 1 0;
+}
+
+/* Balanced with the name column on wide screens (2:1) instead of hugging the
+   right edge; hidden below the tablet breakpoint. */
+.colClientId {
+  @apply gap-1;
+  flex: 1 1 0;
+  min-width: 14rem;
+}
+
+/* CardTable's .cell/.headerCell set display:flex with equal specificity, so a
+   plain .colClientId { mobile:hidden } wins or loses on bundle emission
+   order. The compound selector outranks them deterministically. */
+.colClientId.colClientId {
+  @apply tablet:hidden;
+}
+
+.colActions {
+  width: 3.5rem;
+  flex-shrink: 0;
+  @apply justify-end;
+}
+
+.clientRow {
+  cursor: pointer;
+}
+
+/* Same treatment as .clientHeaderIconWrap on the application edit screen,
+   at list-row size. */
+.clientIconWrap {
+  width: 40px;
+  height: 40px;
+  border-radius: 8px;
+  background: #eef4ff;
+  flex-shrink: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.clientIcon {
+  font-size: 26px;
+  line-height: 1;
+  color: #0969da;
+}
+
+.clientNameBlock {
+  @apply flex min-w-0 flex-col gap-0.5;
+}
+
+.clientName {
+  @apply min-w-0 truncate;
+  margin: 0;
+  font-weight: 600;
+  color: var(--gray-12);
+}
+
+.clientSubtitle {
+  @apply min-w-0 truncate;
+  margin: 0;
+  color: var(--gray-a11);
+}
+
+.clientIdText {
+  @apply min-w-0 truncate;
+  margin: 0;
+  font-family: var(--code-font-family);
+  color: var(--gray-a11);
+}
+
+.compactClientId {
+  @apply hidden tablet:flex items-center gap-1;
+}
+
+.rowActionsButton:not(:hover):not([data-state="open"]) {
+  background-color: transparent;
+}
+
+.listHeader {
+  @apply flex items-center justify-between gap-4;
+}
+
+.listHeader .pageDescription {
+  @apply min-w-0 flex-1;
+}
